+++ backend/judicial-warrant/src/main/java/com/informatique/gov/judicialwarrant/service/impl/CandidateStatusServiceImpl.java
@@ -20,7 +20,7 @@ import lombok.AllArgsConstructor;
 @AllArgsConstructor
 public class CandidateStatusServiceImpl implements EntitledStatusService {
 	private EntitledStatusRepository candidateStatusRepository;
-	private ModelMapper<EntitledStatus, EntitledStatusDto, Byte> candidateStatusMapper;
+	private ModelMapper<EntitledStatus, EntitledStatusDto, Short> candidateStatusMapper;
 
 	/**
 	 *
