+++ backend/judicial-warrant/src/main/java/com/informatique/gov/judicialwarrant/support/modelmpper/EntitledHistoryLogMapper.java
@@ -15,7 +15,7 @@ import lombok.AllArgsConstructor;
 @AllArgsConstructor
 public class EntitledHistoryLogMapper extends AbstractModelMapper<EntitledHistoryLog, EntitledHistoryLogDto, Long> {
 
-	private ModelMapper<EntitledStatus, EntitledStatusDto, Byte> entitledStatusMapper;
+	private ModelMapper<EntitledStatus, EntitledStatusDto, Short> entitledStatusMapper;
 	private ModelMapper<User, UserDto, Integer> userMapper;
 
 	/**
