+++ backend/judicial-warrant/src/main/java/com/informatique/gov/judicialwarrant/support/modelmpper/RequestInternalStatusMapper.java
@@ -7,7 +7,7 @@ import com.informatique.gov.judicialwarrant.rest.dto.RequestStatusDto;
 
 
 @Component
-public class RequestInternalStatusMapper extends AbstractModelMapper<RequestInternalStatus, RequestStatusDto, Byte> {
+public class RequestInternalStatusMapper extends AbstractModelMapper<RequestInternalStatus, RequestStatusDto, Short> {
 
 	/**
 	 *
